requirements.txt
flask
selenium
requests
bs4
pyaml